#252 refactor, create new test folder for tests of calc_world_metrics process
