+++ src/_world/test_world_calc/test_world_calc__char.py
